US-3491 Added socket state check and reconnect
Since the socket server is in the Android activity the disconnection is
the expected behavior. When the activity returns to the foreground the
socket server is available again.

This change introduces a check on the socket writability in the execute
socked command function. If the socket is not writable we will attempt
to reconnect before issuing the command.