US-5457 Fixed proxied caps to allow support for all caps
Problem:
We currently manually create all caps of interest before passing them to a proxied driver.
If the driver supports new caps, we need to modify our driver to add them.

Solution:
This change will avoid having to modify the driver as it will just pass the caps the user has defined.
The following warning will now appear in the proxied driver:
"The following capabilities were provided, but are not recognized by appium: youiEngineAppAddress."
This doesn't cause any problem.